requirements.txt
streamlit
azure-search-documents
openai
python-dotenv
pypdf2